// src/types/futures.d.ts
export type PositionSide = 'BOTH' | 'LONG' | 'SHORT';
export type MarginType = 'ISOLATED' | 'CROSSED';
export type WorkingType = 'MARK_PRICE' | 'CONTRACT_PRICE';
export type FuturesOrderType = 'LIMIT' | 'MARKET' | 'STOP' | 'STOP_MARKET' | 'TAKE_PROFIT' | 'TAKE_PROFIT_MARKET' | 'TRAILING_STOP_MARKET';
export type PriceMatchMode = 'NONE' | 'OPPONENT' | 'OPPONENT_5' | 'OPPONENT_10' | 'OPPONENT_20' | 'QUEUE' | 'QUEUE_5' | 'QUEUE_10' | 'QUEUE_20';
export interface NewFuturesOrderParams<numberType = number> {
    symbol: string;
    side: OrderSide;
    positionSide?: PositionSide;
    type: FuturesOrderType;
    timeInForce?: OrderTimeInForce;
    quantity?: numberType;
    reduceOnly?: BooleanString;
    price?: numberType;
    newClientOrderId?: string;
    stopPrice?: numberType;
    closePosition?: BooleanString;
    activationPrice?: numberType;
    callbackRate?: numberType;
    workingType?: WorkingType;
    priceProtect?: BooleanStringCapitalised;
    newOrderRespType?: OrderResponseType;
    selfTradePreventionMode?: SelfTradePreventionMode;
    priceMatch?: PriceMatchMode;
}
export interface ModifyFuturesOrderParams<numberType = number> {
    orderId?: number;
    origClientOrderId?: string;
    symbol: string;
    side: OrderSide;
    quantity?: numberType;
    price?: numberType;
}
export declare enum EnumPositionMarginChangeType {
    AddPositionMargin = 1,
    ReducePositionMargin = 0
}
export type PositionMarginChangeType = `${EnumPositionMarginChangeType}`;
export type IncomeType = 'TRANSFER' | 'WELCOME_BONUS' | 'REALIZED_PNL' | 'FUNDING_FEE' | 'COMMISSION' | 'INSURANCE_CLEAR';
// ...
export interface GetIncomeHistoryParams {
    symbol?: string;
    incomeType?: IncomeType;
    startTime?: number;
    endTime?: number;
    limit?: number;
    page?: number;
}
export interface IncomeHistory {
    symbol?: string;
    incomeType: IncomeType;
    income: string;
    asset: string;
    time: number;
    info: string;
    tranId: number;
    tradeId: string;
}
// ...
export interface PositionModeResponse {
    dualSidePosition: boolean;
}
export interface MultiAssetModeResponse {
    multiAssetsMargin: boolean;
}
// ...
export interface FuturesAccountBalance {
    accountAlias: string;
    asset: string;
    balance: numberInString;
    crossWalletBalance: numberInString;
    crossUnPnl: numberInString;
    availableBalance: numberInString;
    maxWithdrawAmount: numberInString;
    marginAvailable: boolean;
    updateTime: numberInString;
}
export interface FuturesCoinMAccountBalance {
    accountAlias: string;
    asset: string;
    balance: numberInString;
    withdrawAvailable: numberInString;
    crossWalletBalance: numberInString;
    crossUnPnl: numberInString;
    availableBalance: numberInString;
    updateTime: number;
}
export interface FuturesAccountAsset {
    asset: string;
    walletBalance: numberInString;
    unrealizedProfit: numberInString;
    marginBalance: numberInString;
    maintMargin: numberInString;
    initialMargin: numberInString;
    positionInitialMargin: numberInString;
    openOrderInitialMargin: numberInString;
    maxWithdrawAmount: numberInString;
    crossWalletBalance: numberInString;
    crossUnPnl: numberInString;
    availableBalance: numberInString;
    marginAvailable: boolean;
    updateTime: number;
}
export interface FuturesAccountPosition {
    symbol: string;
    initialMargin: numberInString;
    maintMargin: numberInString;
    unrealizedProfit: numberInString;
    positionInitialMargin: numberInString;
    openOrderInitialMargin: numberInString;
    leverage: numberInString;
    isolated: boolean;
    entryPrice: numberInString;
    maxNotional: numberInString;
    positionSide: PositionSide;
    positionAmt: numberInString;
    notional: numberInString;
    isolatedWallet: numberInString;
    updateTime: number;
    bidNotional: numberInString;
    askNotional: numberInString;
}
// ...
export interface FuturesAccountInformation {
    feeTier: numberInString;
    canTrade: boolean;
    canDeposit: boolean;
    canWithdraw: boolean;
    updateTime: numberInString;
    multiAssetsMargin: boolean;
    totalInitialMargin: numberInString;
    totalMaintMargin: numberInString;
    totalWalletBalance: numberInString;
    totalUnrealizedProfit: numberInString;
    totalMarginBalance: numberInString;
    totalPositionInitialMargin: numberInString;
    totalOpenOrderInitialMargin: numberInString;
    totalCrossWalletBalance: numberInString;
    totalCrossUnPnl: numberInString;
    availableBalance: numberInString;
    maxWithdrawAmount: numberInString;
    assets: FuturesAccountAsset[];
    positions: FuturesAccountPosition[];
}
// ...
export interface SymbolLeverageBracket {
    bracket: number;
    initialLeverage: number;
    notionalCap: number;
    notionalFloor: number;
    maintMarginRatio: number;
    cum: number;
}
export interface SymbolLeverageBracketsResult {
    symbol: string;
    brackets: SymbolLeverageBracket[];
}
export interface UserCommissionRate {
    symbol: string;
    makerCommissionRate: numberInString;
    takerCommissionRate: numberInString;
}
// ...
export interface RebateDataOverview {
    brokerId: string;
    newTraderRebateCommission: numberInString;
    oldTraderRebateCommission: numberInString;
    totalTradeUser: number;
    unit: string;
    totalTradeVol: numberInString;
    totalRebateVol: numberInString;
    time: number;
}
// ...
export interface FuturesTradeHistoryDownloadId {
    avgCostTimestampOfLast30d: number;
    downloadId: string;
}
export interface FuturesTransactionDownloadLink {
    downloadId: string;
    status: 'completed' | 'processing';
    url: string;
    expirationTimestamp: number;
    isExpired: boolean | null;
}
export interface PortfolioMarginProAccountInfo {
    maxWithdrawAmountUSD: string;
    asset: string;
    maxWithdrawAmount: string;
}